pyproject.toml
[tool.poetry]
name = "fast-zero"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.*"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
ignr = "^2.2"
pydantic = {extras = ["email"], version = "^2.7.1"}
sqlalchemy = "^2.0.30"
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"
pyjwt = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
tzdata = "^2024.1"
python-multipart = "^0.0.9"
psycopg = {extras = ["binary"], version = "^3.1.19"}


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-cov = "^5.0.0"
taskipy = "^1.12.2"
ruff = "^0.4.3"
httpx = "^0.27.0"
factory-boy = "^3.3.0"
freezegun = "^1.5.1"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check --diff && ruff check'
format = 'ruff format && ruff check --fix'
run = 'uvicorn fast_zero.app:app --reload'
pre_test = 'task lint'
test = 'pytest -s -x --cov=fast_zero -vv'
post_test = 'coverage html'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"